Split agent restore tests and policies
