+++ cmuxTests/CLINotifyProcessIntegrationRegressionTests.swift
@@ -2,35 +2,6 @@ import XCTest
 import Darwin
 
 final class CLINotifyProcessIntegrationRegressionTests: XCTestCase {
-    private struct ProcessRunResult {
-        let status: Int32
-        let stdout: String
-        let stderr: String
-        let timedOut: Bool
-    }
-
-    private struct GenericHookPersistenceScenario {
-        let agent: String
-        let subcommand: String
-        let sessionId: String
-        let executable: String
-        let launchArguments: [String]
-        let extraEnvironment: [String: String]
-        let expectedArguments: [String]
-        let expectedEnvironment: [String: String]?
-    }
-
-    private final class MockSocketServerState: @unchecked Sendable {
-        private let lock = NSLock()
-        private(set) var commands: [String] = []
-
-        func append(_ command: String) {
-            lock.lock()
-            commands.append(command)
-            lock.unlock()
-        }
-    }
-
     @MainActor
     func testNotifyWithUUIDSurfaceKeepsCallerWorkspaceFallback() throws {
         let cliPath = try bundledCLIPath()
@@ -205,846 +176,4 @@ final class CLINotifyProcessIntegrationRegressionTests: XCTestCase {
         )
     }
 
-    func testGenericHookAgentsPersistSanitizedLaunchCommandsForSessionRestore() throws {
-        let scenarios: [GenericHookPersistenceScenario] = [
-            GenericHookPersistenceScenario(
-                agent: "cursor",
-                subcommand: "prompt-submit",
-                sessionId: "cursor-session-123",
-                executable: "/Users/example/.local/bin/cursor-agent",
-                launchArguments: [
-                    "/Users/example/.local/bin/cursor-agent",
-                    "agent",
-                    "--model",
-                    "gpt-5.4",
-                    "--resume",
-                    "old-chat",
-                    "--workspace",
-                    "/tmp/old repo",
-                    "--sandbox",
-                    "enabled",
-                    "initial prompt should not persist"
-                ],
-                extraEnvironment: [:],
-                expectedArguments: [
-                    "/Users/example/.local/bin/cursor-agent",
-                    "--model",
-                    "gpt-5.4",
-                    "--sandbox",
-                    "enabled"
-                ],
-                expectedEnvironment: nil
-            ),
-            GenericHookPersistenceScenario(
-                agent: "gemini",
-                subcommand: "session-start",
-                sessionId: "gemini-session-123",
-                executable: "/Users/example/.bun/bin/gemini",
-                launchArguments: [
-                    "/Users/example/.bun/bin/gemini",
-                    "--model",
-                    "gemini-2.5-pro",
-                    "--resume",
-                    "old-session",
-                    "--sandbox",
-                    "danger-full-access",
-                    "initial prompt should not persist"
-                ],
-                extraEnvironment: [
-                    "GEMINI_CLI_HOME": "/tmp/gemini home",
-                    "GEMINI_API_KEY": "secret"
-                ],
-                expectedArguments: [
-                    "/Users/example/.bun/bin/gemini",
-                    "--model",
-                    "gemini-2.5-pro",
-                    "--sandbox",
-                    "danger-full-access"
-                ],
-                expectedEnvironment: ["GEMINI_CLI_HOME": "/tmp/gemini home"]
-            ),
-            GenericHookPersistenceScenario(
-                agent: "copilot",
-                subcommand: "session-start",
-                sessionId: "copilot-session-123",
-                executable: "/tmp/cmux-agent-upstreams/copilot-install/bin/copilot",
-                launchArguments: [
-                    "/tmp/cmux-agent-upstreams/copilot-install/bin/copilot",
-                    "--model",
-                    "gpt-5.4",
-                    "--resume=old-session",
-                    "--allow-all-tools",
-                    "-i",
-                    "old prompt",
-                    "initial prompt should not persist"
-                ],
-                extraEnvironment: [
-                    "COPILOT_HOME": "/tmp/copilot home",
-                    "COPILOT_GITHUB_TOKEN": "secret"
-                ],
-                expectedArguments: [
-                    "/tmp/cmux-agent-upstreams/copilot-install/bin/copilot",
-                    "--model",
-                    "gpt-5.4",
-                    "--allow-all-tools"
-                ],
-                expectedEnvironment: ["COPILOT_HOME": "/tmp/copilot home"]
-            ),
-            GenericHookPersistenceScenario(
-                agent: "codebuddy",
-                subcommand: "session-start",
-                sessionId: "codebuddy-session-123",
-                executable: "/Users/example/.npm/bin/codebuddy",
-                launchArguments: [
-                    "/Users/example/.npm/bin/codebuddy",
-                    "--model",
-                    "gpt-5.4",
-                    "--resume",
-                    "old-session",
-                    "--permission-mode",
-                    "plan",
-                    "--worktree",
-                    "scratch",
-                    "initial prompt should not persist"
-                ],
-                extraEnvironment: [
-                    "CODEBUDDY_CONFIG_DIR": "/tmp/codebuddy config",
-                    "CODEBUDDY_API_KEY": "secret"
-                ],
-                expectedArguments: [
-                    "/Users/example/.npm/bin/codebuddy",
-                    "--model",
-                    "gpt-5.4",
-                    "--permission-mode",
-                    "plan"
-                ],
-                expectedEnvironment: ["CODEBUDDY_CONFIG_DIR": "/tmp/codebuddy config"]
-            ),
-            GenericHookPersistenceScenario(
-                agent: "factory",
-                subcommand: "session-start",
-                sessionId: "factory-session-123",
-                executable: "/Users/example/.npm/bin/droid",
-                launchArguments: [
-                    "/Users/example/.npm/bin/droid",
-                    "--resume",
-                    "old-session",
-                    "--cwd",
-                    "/tmp/factory repo",
-                    "--append-system-prompt",
-                    "be terse",
-                    "initial prompt should not persist"
-                ],
-                extraEnvironment: [
-                    "FACTORY_API_KEY": "secret"
-                ],
-                expectedArguments: [
-                    "/Users/example/.npm/bin/droid",
-                    "--cwd",
-                    "/tmp/factory repo",
-                    "--append-system-prompt",
-                    "be terse"
-                ],
-                expectedEnvironment: nil
-            ),
-            GenericHookPersistenceScenario(
-                agent: "qoder",
-                subcommand: "session-start",
-                sessionId: "qoder-session-123",
-                executable: "/Users/example/.npm/bin/qodercli",
-                launchArguments: [
-                    "/Users/example/.npm/bin/qodercli",
-                    "--model",
-                    "gemini-2.5-pro",
-                    "--resume",
-                    "old-session",
-                    "--permission-mode",
-                    "plan",
-                    "--workspace",
-                    "/tmp/qoder repo",
-                    "initial prompt should not persist"
-                ],
-                extraEnvironment: [
-                    "QODER_CONFIG_DIR": "/tmp/qoder config",
-                    "GEMINI_API_KEY": "secret"
-                ],
-                expectedArguments: [
-                    "/Users/example/.npm/bin/qodercli",
-                    "--model",
-                    "gemini-2.5-pro",
-                    "--permission-mode",
-                    "plan",
-                    "--workspace",
-                    "/tmp/qoder repo"
-                ],
-                expectedEnvironment: ["QODER_CONFIG_DIR": "/tmp/qoder config"]
-            ),
-        ]
-
-        for scenario in scenarios {
-            try XCTContext.runActivity(named: scenario.agent) { _ in
-                try runGenericHookPersistenceScenario(scenario)
-            }
-        }
-    }
-
-    func testRovoDevPromptSubmitInfersSessionIdFromWorkspaceMetadataAndPersistsLaunchCommand() throws {
-        let cliPath = try bundledCLIPath()
-        let socketPath = makeSocketPath("rovo-infer")
-        let listenerFD = try bindUnixSocket(at: socketPath)
-        let state = MockSocketServerState()
-        let root = FileManager.default.temporaryDirectory
-            .appendingPathComponent("cmux-rovo-hook-\(UUID().uuidString)", isDirectory: true)
-        let workspace = root.appendingPathComponent("repo", isDirectory: true)
-        let sessionsRoot = root.appendingPathComponent("sessions", isDirectory: true)
-        let workspaceId = "55555555-5555-5555-5555-555555555555"
-        let surfaceId = "66666666-6666-6666-6666-666666666666"
-        let olderSessionId = "rovo-older-session"
-        let newestSessionId = "rovo-newest-session"
-
-        try FileManager.default.createDirectory(at: workspace, withIntermediateDirectories: true)
-        try writeRovoDevSessionMetadata(
-            sessionsRoot: sessionsRoot,
-            sessionId: olderSessionId,
-            workspacePath: workspace.path,
-            modified: Date(timeIntervalSince1970: 100)
-        )
-        try writeRovoDevSessionMetadata(
-            sessionsRoot: sessionsRoot,
-            sessionId: newestSessionId,
-            workspacePath: workspace.path,
-            modified: Date(timeIntervalSince1970: 200)
-        )
-        defer {
-            Darwin.close(listenerFD)
-            unlink(socketPath)
-            try? FileManager.default.removeItem(at: root)
-        }
-
-        let serverHandled = startMockServer(listenerFD: listenerFD, state: state) { line in
-            guard let payload = self.jsonObject(line) else {
-                return "OK"
-            }
-            guard let id = payload["id"] as? String, let method = payload["method"] as? String else {
-                return self.malformedRequestResponse(id: payload["id"] as? String, raw: line)
-            }
-            if method == "surface.list" {
-                return self.surfaceListResponse(id: id, surfaceId: surfaceId)
-            }
-            return self.v2Response(id: id, ok: false, error: ["code": "unrecognized_method", "message": "unexpected method: \(method)"])
-        }
-
-        var environment = ProcessInfo.processInfo.environment
-        environment["CMUX_SOCKET_PATH"] = socketPath
-        environment["CMUX_WORKSPACE_ID"] = workspaceId
-        environment["CMUX_SURFACE_ID"] = surfaceId
-        environment["CMUX_AGENT_HOOK_STATE_DIR"] = root.path
-        environment["CMUX_ROVODEV_SESSIONS_DIR"] = sessionsRoot.path
-        environment["CMUX_AGENT_LAUNCH_KIND"] = "rovodev"
-        environment["CMUX_AGENT_LAUNCH_EXECUTABLE"] = "/usr/local/bin/acli"
-        environment["CMUX_AGENT_LAUNCH_ARGV_B64"] = base64NULSeparated([
-            "/usr/local/bin/acli",
-            "rovodev",
-            "run",
-            "--restore",
-            olderSessionId,
-            "--yolo",
-            "prompt that should not persist",
-        ])
-        environment["CMUX_AGENT_LAUNCH_CWD"] = workspace.path
-        environment["CMUX_CLI_SENTRY_DISABLED"] = "1"
-
-        let result = runProcess(
-            executablePath: cliPath,
-            arguments: ["hooks", "rovodev", "prompt-submit"],
-            environment: environment,
-            standardInput: #"{"cwd":"\#(workspace.path)","hook_event_name":"on_tool_permission"}"#,
-            timeout: 5
-        )
-
-        wait(for: [serverHandled], timeout: 5)
-        XCTAssertFalse(result.timedOut, result.stderr)
-        XCTAssertEqual(result.status, 0, result.stderr)
-        XCTAssertEqual(result.stdout, "{}\n")
-
-        let storeURL = root.appendingPathComponent("rovodev-hook-sessions.json", isDirectory: false)
-        let json = try XCTUnwrap(JSONSerialization.jsonObject(with: Data(contentsOf: storeURL)) as? [String: Any])
-        let sessions = try XCTUnwrap(json["sessions"] as? [String: Any])
-        let session = try XCTUnwrap(sessions[newestSessionId] as? [String: Any])
-        XCTAssertNil(sessions[olderSessionId] as? [String: Any])
-        XCTAssertEqual(session["workspaceId"] as? String, workspaceId)
-        XCTAssertEqual(session["surfaceId"] as? String, surfaceId)
-        XCTAssertEqual(session["cwd"] as? String, workspace.path)
-
-        let launchCommand = try XCTUnwrap(session["launchCommand"] as? [String: Any])
-        XCTAssertEqual(launchCommand["launcher"] as? String, "rovodev")
-        XCTAssertEqual(launchCommand["executablePath"] as? String, "/usr/local/bin/acli")
-        XCTAssertEqual(
-            launchCommand["arguments"] as? [String],
-            ["/usr/local/bin/acli", "rovodev", "run", "--yolo"]
-        )
-        XCTAssertTrue(
-            state.commands.contains { $0.contains("set_status rovodev Running") && $0.contains("--tab=\(workspaceId)") },
-            "Expected Rovo Dev prompt status to target current workspace, saw \(state.commands)"
-        )
-    }
-
-    func testRovoDevPromptSubmitReadsConfiguredPersistenceDirWithCommentsAndHashPath() throws {
-        let cliPath = try bundledCLIPath()
-        let socketPath = makeSocketPath("rovo-config")
-        let listenerFD = try bindUnixSocket(at: socketPath)
-        let state = MockSocketServerState()
-        let root = FileManager.default.temporaryDirectory
-            .appendingPathComponent("cmux-rovo-config-hook-\(UUID().uuidString)", isDirectory: true)
-        let workspace = root.appendingPathComponent("repo", isDirectory: true)
-        let sessionsRoot = root.appendingPathComponent("sessions#hash", isDirectory: true)
-        let configDir = root.appendingPathComponent(".rovodev", isDirectory: true)
-        let workspaceId = "55555555-5555-5555-5555-555555555555"
-        let surfaceId = "66666666-6666-6666-6666-666666666666"
-        let sessionId = "rovo-config-session"
-
-        try FileManager.default.createDirectory(at: workspace, withIntermediateDirectories: true)
-        try FileManager.default.createDirectory(at: configDir, withIntermediateDirectories: true)
-        try writeRovoDevSessionMetadata(
-            sessionsRoot: sessionsRoot,
-            sessionId: sessionId,
-            workspacePath: workspace.path,
-            modified: Date(timeIntervalSince1970: 300)
-        )
-        let config = [
-            "sessions:",
-            "  # top-level comments inside sessions should not end the block",
-            "  nested:",
-            "    persistenceDir: /tmp/wrong",
-            "  persistenceDir: '~/sessions#hash'",
-            "other: true",
-        ].joined(separator: "\r\n")
-        try config.write(
-            to: configDir.appendingPathComponent("config.yml", isDirectory: false),
-            atomically: true,
-            encoding: .utf8
-        )
-        defer {
-            Darwin.close(listenerFD)
-            unlink(socketPath)
-            try? FileManager.default.removeItem(at: root)
-        }
-
-        let serverHandled = startMockServer(listenerFD: listenerFD, state: state) { line in
-            guard let payload = self.jsonObject(line) else {
-                return "OK"
-            }
-            guard let id = payload["id"] as? String, let method = payload["method"] as? String else {
-                return self.malformedRequestResponse(id: payload["id"] as? String, raw: line)
-            }
-            if method == "surface.list" {
-                return self.surfaceListResponse(id: id, surfaceId: surfaceId)
-            }
-            if method == "feed.push" {
-                return self.v2Response(id: id, ok: true, result: [:])
-            }
-            return self.v2Response(id: id, ok: false, error: ["code": "unrecognized_method", "message": "unexpected method: \(method)"])
-        }
-
-        let environment: [String: String] = [
-            "HOME": root.path,
-            "PATH": "/usr/bin:/bin:/usr/sbin:/sbin",
-            "CMUX_SOCKET_PATH": socketPath,
-            "CMUX_WORKSPACE_ID": workspaceId,
-            "CMUX_SURFACE_ID": surfaceId,
-            "CMUX_AGENT_HOOK_STATE_DIR": root.path,
-            "CMUX_AGENT_LAUNCH_KIND": "rovodev",
-            "CMUX_AGENT_LAUNCH_EXECUTABLE": "/usr/local/bin/acli",
-            "CMUX_AGENT_LAUNCH_ARGV_B64": base64NULSeparated([
-                "/usr/local/bin/acli",
-                "rovodev",
-                "run",
-                "--restore",
-                sessionId,
-                "--yolo",
-            ]),
-            "CMUX_AGENT_LAUNCH_CWD": workspace.path,
-            "CMUX_CLI_SENTRY_DISABLED": "1",
-        ]
-
-        let result = runProcess(
-            executablePath: cliPath,
-            arguments: ["hooks", "rovodev", "prompt-submit"],
-            environment: environment,
-            standardInput: #"{"cwd":"\#(workspace.path)","hook_event_name":"on_tool_permission"}"#,
-            timeout: 5
-        )
-
-        wait(for: [serverHandled], timeout: 5)
-        XCTAssertFalse(result.timedOut, result.stderr)
-        XCTAssertEqual(result.status, 0, result.stderr)
-        XCTAssertEqual(result.stdout, "{}\n")
-
-        let storeURL = root.appendingPathComponent("rovodev-hook-sessions.json", isDirectory: false)
-        let json = try XCTUnwrap(JSONSerialization.jsonObject(with: Data(contentsOf: storeURL)) as? [String: Any])
-        let sessions = try XCTUnwrap(json["sessions"] as? [String: Any])
-        XCTAssertNotNil(sessions[sessionId] as? [String: Any])
-    }
-
-    func testRovoDevPromptSubmitWithoutCwdDoesNotInferUnrelatedSession() throws {
-        let cliPath = try bundledCLIPath()
-        let socketPath = makeSocketPath("rovo-nocwd")
-        let listenerFD = try bindUnixSocket(at: socketPath)
-        let root = FileManager.default.temporaryDirectory
-            .appendingPathComponent("cmux-rovo-nocwd-hook-\(UUID().uuidString)", isDirectory: true)
-        let sessionsRoot = root.appendingPathComponent("sessions", isDirectory: true)
-        let workspaceId = "55555555-5555-5555-5555-555555555555"
-        let surfaceId = "66666666-6666-6666-6666-666666666666"
-        let unrelatedSessionId = "rovo-unrelated-session"
-
-        try writeRovoDevSessionMetadata(
-            sessionsRoot: sessionsRoot,
-            sessionId: unrelatedSessionId,
-            workspacePath: "/tmp/unrelated",
-            modified: Date(timeIntervalSince1970: 300)
-        )
-        defer {
-            Darwin.close(listenerFD)
-            unlink(socketPath)
-            try? FileManager.default.removeItem(at: root)
-        }
-
-        let state = MockSocketServerState()
-        let serverHandled = startMockServer(listenerFD: listenerFD, state: state) { line in
-            guard let payload = self.jsonObject(line) else {
-                return "OK"
-            }
-            guard let id = payload["id"] as? String, let method = payload["method"] as? String else {
-                return self.malformedRequestResponse(id: payload["id"] as? String, raw: line)
-            }
-            if method == "surface.list" {
-                return self.surfaceListResponse(id: id, surfaceId: surfaceId)
-            }
-            if method == "feed.push" {
-                return self.v2Response(id: id, ok: true, result: [:])
-            }
-            return self.v2Response(id: id, ok: false, error: ["code": "unrecognized_method", "message": "unexpected method: \(method)"])
-        }
-
-        let environment: [String: String] = [
-            "HOME": root.path,
-            "PATH": "/usr/bin:/bin:/usr/sbin:/sbin",
-            "CMUX_SOCKET_PATH": socketPath,
-            "CMUX_WORKSPACE_ID": workspaceId,
-            "CMUX_SURFACE_ID": surfaceId,
-            "CMUX_AGENT_HOOK_STATE_DIR": root.path,
-            "CMUX_ROVODEV_SESSIONS_DIR": sessionsRoot.path,
-            "CMUX_CLI_SENTRY_DISABLED": "1",
-        ]
-
-        let result = runProcess(
-            executablePath: cliPath,
-            arguments: ["hooks", "rovodev", "prompt-submit"],
-            environment: environment,
-            standardInput: #"{"hook_event_name":"on_tool_permission"}"#,
-            timeout: 5
-        )
-
-        wait(for: [serverHandled], timeout: 5)
-        XCTAssertFalse(result.timedOut, result.stderr)
-        XCTAssertEqual(result.status, 0, result.stderr)
-        XCTAssertEqual(result.stdout, "{}\n")
-
-        let storeURL = root.appendingPathComponent("rovodev-hook-sessions.json", isDirectory: false)
-        if let data = try? Data(contentsOf: storeURL),
-           let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any],
-           let sessions = json["sessions"] as? [String: Any] {
-            XCTAssertNil(sessions[unrelatedSessionId] as? [String: Any])
-        }
-    }
-
-    func testLegacyCodexHookAliasReturnsJSONWithoutHelpAndPersistsSession() throws {
-        let cliPath = try bundledCLIPath()
-        let socketPath = makeSocketPath("legacy-codex")
-        let listenerFD = try bindUnixSocket(at: socketPath)
-        let state = MockSocketServerState()
-        let root = FileManager.default.temporaryDirectory
-            .appendingPathComponent("cmux-legacy-codex-hook-\(UUID().uuidString)", isDirectory: true)
-        let workspaceId = "11111111-1111-1111-1111-111111111111"
-        let surfaceId = "22222222-2222-2222-2222-222222222222"
-
-        try FileManager.default.createDirectory(at: root, withIntermediateDirectories: true)
-        defer {
-            Darwin.close(listenerFD)
-            unlink(socketPath)
-            try? FileManager.default.removeItem(at: root)
-        }
-
-        let serverHandled = startMockServer(listenerFD: listenerFD, state: state) { line in
-            guard let payload = self.jsonObject(line) else {
-                return line.trimmingCharacters(in: .whitespacesAndNewlines).hasPrefix("{")
-                    ? self.malformedRequestResponse(raw: line)
-                    : "OK"
-            }
-            guard let id = payload["id"] as? String, let method = payload["method"] as? String else {
-                return self.malformedRequestResponse(id: payload["id"] as? String, raw: line)
-            }
-            if method == "surface.list" { return self.surfaceListResponse(id: id, surfaceId: surfaceId) }
-            return self.v2Response(id: id, ok: false, error: ["code": "unrecognized_method", "message": "unexpected method: \(method)"])
-        }
-
-        var environment = ProcessInfo.processInfo.environment
-        environment["CMUX_SOCKET_PATH"] = socketPath
-        environment["CMUX_WORKSPACE_ID"] = workspaceId
-        environment["CMUX_SURFACE_ID"] = surfaceId
-        environment["CMUX_AGENT_HOOK_STATE_DIR"] = root.path
-        environment["CMUX_AGENT_LAUNCH_KIND"] = "codex"
-        environment["CMUX_AGENT_LAUNCH_EXECUTABLE"] = "/usr/local/bin/codex"
-        environment["CMUX_AGENT_LAUNCH_ARGV_B64"] = base64NULSeparated(["/usr/local/bin/codex", "--model", "gpt-5.4"])
-        environment["CMUX_AGENT_LAUNCH_CWD"] = "/tmp/repo"
-        environment["CODEX_HOME"] = root.appendingPathComponent("codex-home", isDirectory: true).path
-        environment["CMUX_CLI_SENTRY_DISABLED"] = "1"
-
-        let result = runProcess(
-            executablePath: cliPath,
-            arguments: ["codex-hook", "session-start"],
-            environment: environment,
-            timeout: 5
-        )
-
-        wait(for: [serverHandled], timeout: 5)
-        XCTAssertFalse(result.timedOut, result.stderr)
-        XCTAssertEqual(result.status, 0, result.stderr)
-        XCTAssertEqual(result.stdout, "{}\n")
-        XCTAssertFalse(result.stdout.contains("Usage:"), result.stdout)
-        XCTAssertFalse(result.stderr.contains("Usage:"), result.stderr)
-
-        let storeURL = root.appendingPathComponent("codex-hook-sessions.json", isDirectory: false)
-        let json = try XCTUnwrap(JSONSerialization.jsonObject(with: Data(contentsOf: storeURL)) as? [String: Any])
-        let sessions = try XCTUnwrap(json["sessions"] as? [String: Any])
-        let session = try XCTUnwrap(sessions[surfaceId] as? [String: Any])
-        XCTAssertEqual(session["workspaceId"] as? String, workspaceId)
-        XCTAssertEqual(session["surfaceId"] as? String, surfaceId)
-        XCTAssertNotNil(session["launchCommand"] as? [String: Any])
-    }
-
-    func testLegacyFeedHookAliasReturnsJSONWithoutHelpOutsideCmuxTerminal() throws {
-        let cliPath = try bundledCLIPath()
-        var environment = ProcessInfo.processInfo.environment
-        environment["CMUX_SOCKET_PATH"] = makeSocketPath("legacy-feed")
-        environment.removeValue(forKey: "CMUX_WORKSPACE_ID")
-        environment.removeValue(forKey: "CMUX_SURFACE_ID")
-        environment["CMUX_CLI_SENTRY_DISABLED"] = "1"
-
-        let result = runProcess(
-            executablePath: cliPath,
-            arguments: ["feed-hook", "--source", "codex"],
-            environment: environment,
-            standardInput: #"{"hook_event_name":"UserPromptSubmit","session_id":"legacy-feed"}"#,
-            timeout: 5
-        )
-
-        XCTAssertFalse(result.timedOut, result.stderr)
-        XCTAssertEqual(result.status, 0, result.stderr)
-        XCTAssertEqual(result.stdout, "{}\n")
-        XCTAssertFalse(result.stdout.contains("Usage:"), result.stdout)
-        XCTAssertFalse(result.stderr.contains("Usage:"), result.stderr)
-    }
-
-    private func runGenericHookPersistenceScenario(_ scenario: GenericHookPersistenceScenario) throws {
-        let cliPath = try bundledCLIPath()
-        let socketPath = makeSocketPath("hook-\(scenario.agent)")
-        let listenerFD = try bindUnixSocket(at: socketPath)
-        let state = MockSocketServerState()
-        let root = FileManager.default.temporaryDirectory
-            .appendingPathComponent("cmux-\(scenario.agent)-hook-\(UUID().uuidString)", isDirectory: true)
-        let workspace = root.appendingPathComponent("repo", isDirectory: true)
-        let workspaceId = "11111111-1111-1111-1111-111111111111"
-        let surfaceId = "22222222-2222-2222-2222-222222222222"
-
-        try FileManager.default.createDirectory(at: workspace, withIntermediateDirectories: true)
-        defer {
-            Darwin.close(listenerFD)
-            unlink(socketPath)
-            try? FileManager.default.removeItem(at: root)
-        }
-
-        let serverHandled = startMockServer(listenerFD: listenerFD, state: state) { line in
-            guard let payload = self.jsonObject(line) else {
-                return "OK"
-            }
-            guard let id = payload["id"] as? String, let method = payload["method"] as? String else {
-                return self.malformedRequestResponse(id: payload["id"] as? String, raw: line)
-            }
-            switch method {
-            case "surface.list":
-                return self.surfaceListResponse(id: id, surfaceId: surfaceId)
-            case "feed.push":
-                return self.v2Response(id: id, ok: true, result: [:])
-            default:
-                return self.v2Response(id: id, ok: false, error: ["code": "unrecognized_method", "message": "unexpected method: \(method)"])
-            }
-        }
-
-        var environment: [String: String] = [
-            "HOME": root.path,
-            "PATH": "/usr/bin:/bin:/usr/sbin:/sbin",
-            "PWD": workspace.path,
-            "CMUX_SOCKET_PATH": socketPath,
-            "CMUX_WORKSPACE_ID": workspaceId,
-            "CMUX_SURFACE_ID": surfaceId,
-            "CMUX_AGENT_HOOK_STATE_DIR": root.path,
-            "CMUX_AGENT_LAUNCH_KIND": scenario.agent,
-            "CMUX_AGENT_LAUNCH_EXECUTABLE": scenario.executable,
-            "CMUX_AGENT_LAUNCH_ARGV_B64": base64NULSeparated(scenario.launchArguments),
-            "CMUX_AGENT_LAUNCH_CWD": workspace.path,
-            "CMUX_CLI_SENTRY_DISABLED": "1",
-        ]
-        for (key, value) in scenario.extraEnvironment {
-            environment[key] = value
-        }
-
-        let result = runProcess(
-            executablePath: cliPath,
-            arguments: ["hooks", scenario.agent, scenario.subcommand],
-            environment: environment,
-            standardInput: #"{"session_id":"\#(scenario.sessionId)","cwd":"\#(workspace.path)","hook_event_name":"SessionStart"}"#,
-            timeout: 5
-        )
-
-        wait(for: [serverHandled], timeout: 5)
-        XCTAssertFalse(result.timedOut, result.stderr)
-        XCTAssertEqual(result.status, 0, result.stderr)
-        XCTAssertEqual(result.stdout, "{}\n")
-
-        let storeURL = root.appendingPathComponent("\(scenario.agent)-hook-sessions.json", isDirectory: false)
-        let json = try XCTUnwrap(JSONSerialization.jsonObject(with: Data(contentsOf: storeURL)) as? [String: Any])
-        let sessions = try XCTUnwrap(json["sessions"] as? [String: Any])
-        let session = try XCTUnwrap(sessions[scenario.sessionId] as? [String: Any])
-        XCTAssertEqual(session["workspaceId"] as? String, workspaceId)
-        XCTAssertEqual(session["surfaceId"] as? String, surfaceId)
-        XCTAssertEqual(session["cwd"] as? String, workspace.path)
-
-        let launchCommand = try XCTUnwrap(session["launchCommand"] as? [String: Any])
-        XCTAssertEqual(launchCommand["launcher"] as? String, scenario.agent)
-        XCTAssertEqual(launchCommand["executablePath"] as? String, scenario.executable)
-        XCTAssertEqual(launchCommand["arguments"] as? [String], scenario.expectedArguments)
-        XCTAssertEqual(launchCommand["workingDirectory"] as? String, workspace.path)
-        XCTAssertEqual(launchCommand["environment"] as? [String: String], scenario.expectedEnvironment)
-    }
-
-    private func bundledCLIPath() throws -> String {
-        let fileManager = FileManager.default
-        let appBundleURL = Bundle(for: Self.self)
-            .bundleURL
-            .deletingLastPathComponent()
-            .deletingLastPathComponent()
-            .deletingLastPathComponent()
-        let enumerator = fileManager.enumerator(
-            at: appBundleURL,
-            includingPropertiesForKeys: nil,
-            options: [.skipsHiddenFiles]
-        )
-
-        while let item = enumerator?.nextObject() as? URL {
-            guard item.lastPathComponent == "cmux",
-                  item.path.contains(".app/Contents/Resources/bin/cmux") else {
-                continue
-            }
-            return item.path
-        }
-
-        throw XCTSkip("Bundled cmux CLI not found in \(appBundleURL.path)")
-    }
-
-    private func makeSocketPath(_ name: String) -> String {
-        let shortID = UUID().uuidString.replacingOccurrences(of: "-", with: "").prefix(8)
-        return URL(fileURLWithPath: NSTemporaryDirectory())
-            .appendingPathComponent("cli-\(name.prefix(6))-\(shortID).sock")
-            .path
-    }
-
-    private func bindUnixSocket(at path: String) throws -> Int32 {
-        unlink(path)
-        let fd = Darwin.socket(AF_UNIX, SOCK_STREAM, 0)
-        XCTAssertGreaterThanOrEqual(fd, 0)
-
-        var addr = sockaddr_un()
-        addr.sun_family = sa_family_t(AF_UNIX)
-        let maxPathLength = MemoryLayout.size(ofValue: addr.sun_path)
-        let utf8 = Array(path.utf8)
-        XCTAssertLessThan(utf8.count, maxPathLength)
-        _ = withUnsafeMutablePointer(to: &addr.sun_path) { pointer in
-            pointer.withMemoryRebound(to: CChar.self, capacity: maxPathLength) { buffer in
-                for index in 0..<utf8.count {
-                    buffer[index] = CChar(bitPattern: utf8[index])
-                }
-                buffer[utf8.count] = 0
-            }
-        }
-
-        let bindResult = withUnsafePointer(to: &addr) { pointer in
-            pointer.withMemoryRebound(to: sockaddr.self, capacity: 1) { sockaddrPtr in
-                Darwin.bind(fd, sockaddrPtr, socklen_t(MemoryLayout<sockaddr_un>.size))
-            }
-        }
-        XCTAssertEqual(bindResult, 0)
-        XCTAssertEqual(Darwin.listen(fd, 1), 0)
-        return fd
-    }
-
-    private func startMockServer(
-        listenerFD: Int32,
-        state: MockSocketServerState,
-        handler: @escaping @Sendable (String) -> String
-    ) -> XCTestExpectation {
-        let handled = expectation(description: "cli mock socket handled")
-        DispatchQueue.global(qos: .userInitiated).async {
-            var clientAddr = sockaddr_un()
-            var clientAddrLen = socklen_t(MemoryLayout<sockaddr_un>.size)
-            let clientFD = withUnsafeMutablePointer(to: &clientAddr) { ptr in
-                ptr.withMemoryRebound(to: sockaddr.self, capacity: 1) { sockaddrPtr in
-                    Darwin.accept(listenerFD, sockaddrPtr, &clientAddrLen)
-                }
-            }
-            guard clientFD >= 0 else {
-                handled.fulfill()
-                return
-            }
-            defer {
-                Darwin.close(clientFD)
-                handled.fulfill()
-            }
-
-            var pending = Data()
-            var buffer = [UInt8](repeating: 0, count: 4096)
-            while true {
-                let count = Darwin.read(clientFD, &buffer, buffer.count)
-                if count < 0 {
-                    if errno == EINTR { continue }
-                    return
-                }
-                if count == 0 { return }
-                pending.append(buffer, count: count)
-
-                while let newlineRange = pending.firstRange(of: Data([0x0A])) {
-                    let lineData = pending.subdata(in: 0..<newlineRange.lowerBound)
-                    pending.removeSubrange(0...newlineRange.lowerBound)
-                    guard let line = String(data: lineData, encoding: .utf8) else { continue }
-                    state.append(line)
-                    let response = handler(line) + "\n"
-                    _ = response.withCString { ptr in
-                        Darwin.write(clientFD, ptr, strlen(ptr))
-                    }
-                }
-            }
-        }
-        return handled
-    }
-
-    private func v2Response(
-        id: String,
-        ok: Bool,
-        result: [String: Any]? = nil,
-        error: [String: Any]? = nil
-    ) -> String {
-        var payload: [String: Any] = ["id": id, "ok": ok]
-        if let result { payload["result"] = result }
-        if let error { payload["error"] = error }
-        let data = try? JSONSerialization.data(withJSONObject: payload, options: [])
-        return String(data: data ?? Data("{}".utf8), encoding: .utf8) ?? "{}"
-    }
-
-    private func malformedRequestResponse(id: String? = nil, raw: String) -> String {
-        v2Response(
-            id: id ?? "unknown",
-            ok: false,
-            error: ["code": "malformed_request", "message": "invalid or non-JSON payload", "raw": raw]
-        )
-    }
-
-    private func surfaceListResponse(id: String, surfaceId: String) -> String {
-        v2Response(
-            id: id,
-            ok: true,
-            result: ["surfaces": [["id": surfaceId, "ref": "surface:1", "focused": true]]]
-        )
-    }
-
-    private func jsonObject(_ line: String) -> [String: Any]? {
-        guard let data = line.data(using: .utf8) else { return nil }
-        return try? JSONSerialization.jsonObject(with: data, options: []) as? [String: Any]
-    }
-
-    private func base64NULSeparated(_ values: [String]) -> String {
-        var data = Data()
-        for value in values {
-            data.append(contentsOf: value.utf8)
-            data.append(0)
-        }
-        return data.base64EncodedString()
-    }
-
-    private func writeRovoDevSessionMetadata(
-        sessionsRoot: URL,
-        sessionId: String,
-        workspacePath: String,
-        modified: Date
-    ) throws {
-        let sessionURL = sessionsRoot.appendingPathComponent(sessionId, isDirectory: true)
-        try FileManager.default.createDirectory(at: sessionURL, withIntermediateDirectories: true)
-        let metadataURL = sessionURL.appendingPathComponent("metadata.json", isDirectory: false)
-        let metadata = [
-            "title": "Rovo Dev session",
-            "workspace_path": workspacePath,
-        ]
-        let data = try JSONSerialization.data(withJSONObject: metadata, options: [.prettyPrinted])
-        try data.write(to: metadataURL, options: .atomic)
-        try FileManager.default.setAttributes([.modificationDate: modified], ofItemAtPath: metadataURL.path)
-    }
-
-    private func runProcess(
-        executablePath: String,
-        arguments: [String],
-        environment: [String: String],
-        standardInput: String? = nil,
-        timeout: TimeInterval
-    ) -> ProcessRunResult {
-        let process = Process()
-        let stdoutPipe = Pipe()
-        let stderrPipe = Pipe()
-        let stdinPipe = standardInput == nil ? nil : Pipe()
-        process.executableURL = URL(fileURLWithPath: executablePath)
-        process.arguments = arguments
-        process.environment = environment
-        process.standardInput = stdinPipe ?? FileHandle.nullDevice
-        process.standardOutput = stdoutPipe
-        process.standardError = stderrPipe
-
-        do {
-            try process.run()
-        } catch {
-            return ProcessRunResult(status: -1, stdout: "", stderr: String(describing: error), timedOut: false)
-        }
-        if let standardInput, let stdinPipe {
-            stdinPipe.fileHandleForWriting.write(Data(standardInput.utf8))
-            try? stdinPipe.fileHandleForWriting.close()
-        }
-
-        let exitSignal = DispatchSemaphore(value: 0)
-        DispatchQueue.global(qos: .userInitiated).async {
-            process.waitUntilExit()
-            exitSignal.signal()
-        }
-
-        let timedOut = exitSignal.wait(timeout: .now() + timeout) == .timedOut
-        if timedOut {
-            process.terminate()
-            _ = exitSignal.wait(timeout: .now() + 1)
-        }
-
-        let stdout = String(data: stdoutPipe.fileHandleForReading.readDataToEndOfFile(), encoding: .utf8) ?? ""
-        let stderr = String(data: stderrPipe.fileHandleForReading.readDataToEndOfFile(), encoding: .utf8) ?? ""
-        return ProcessRunResult(
-            status: process.terminationStatus,
-            stdout: stdout,
-            stderr: stderr,
-            timedOut: timedOut
-        )
-    }
 }
